license_tools: Split __init__
This should help with future maintenance
