+++ license_tools/__init__.py
@@ -22,553 +22,23 @@ See README.md for detail and documentation
 """
 
 import argparse
-import datetime
-import enum
 import functools
 import json
 import logging
-import os
 import pathlib
-import re
-import subprocess
 import sys
-from collections import namedtuple
 from copy import copy
-from operator import attrgetter
 from typing import Dict
-import jinja2
+
+from .style import Style
+from .utils import DateUtils, FileFilter
+from .license import License, LICENSES
+from .git import GitRepo
+from .header import Header, ParsedHeader, Author, Title
 
 BASE_DIR = pathlib.Path(__file__).parent
 CW_DIR = pathlib.Path.cwd()
-SPDX_LICENSES = list(BASE_DIR.glob('spdx_licenses/*.spdx'))
-OTHER_LICENSES = list(BASE_DIR.glob('other_licenses/*.license'))
-LICENSES = {license_file.stem: license_file for license_file in SPDX_LICENSES + OTHER_LICENSES}
 LICENSE_JSON = '.license-tools-config.json'
-
-
-class DateUtils:
-    """Utilities for date functions used in this package"""
-    _current_year = None
-
-    @staticmethod
-    def current_year():
-        """
-        Returns the current year
-
-        Override using the LICTOOLS_OVERRIDE_YEAR env variable
-        """
-        if DateUtils._current_year is None:
-            year = os.getenv('LICTOOLS_OVERRIDE_YEAR', datetime.date.today().year)
-            DateUtils._current_year = int(year)
-        return DateUtils._current_year
-
-    @staticmethod
-    def parse_git_date(git_date: str) -> datetime.datetime:
-        """
-        Returns a datetime parsed from a git date string
-        as documented at https://git-scm.com/docs/git-commit/2.24.0#_date_formats
-        """
-        # Git internal format
-        try:
-            if git_date.startswith('@'):
-                unix_timestamp, tz_offset = git_date[1:].split(' ', maxsplit=1)
-                unix_timestamp = int(unix_timestamp)
-                timezone = datetime.timezone(datetime.datetime.strptime(tz_offset, '%z').utcoffset())
-                return datetime.datetime.fromtimestamp(unix_timestamp, tz=timezone)
-        except:  # pylint: disable=bare-except
-            pass
-
-        # RFC 2822
-        try:
-            from email.utils import parsedate_to_datetime  # pylint: disable=import-outside-toplevel
-            author_date = parsedate_to_datetime(git_date)
-            if author_date:
-                return author_date
-        except:  # pylint: disable=bare-except
-            pass
-
-        # ISO 8601
-        try:
-            return datetime.datetime.fromisoformat(git_date)
-        except:  # pylint: disable=bare-except
-            pass
-
-        # plain formats
-        for format in ['%Y.%m.%d', '%m/%d/%Y', '%d.%m.%Y']:
-            try:
-                return datetime.datetime.strptime(git_date, format)
-            except:  # pylint: disable=bare-except
-                pass
-
-        raise RuntimeError(f"Not a supported git date format: {git_date}")
-
-
-class Style(enum.Enum):
-    """Enumerates the different known comment styles"""
-    UNKNOWN = 1
-    C_STYLE = 2
-    POUND_STYLE = 3
-    DOCSTRING_STYLE = 4
-    XML_STYLE = 5
-    BATCH_STYLE = 6
-    SLASH_STYLE = 7
-    DASH_STYLE = 8
-
-    @classmethod
-    def set_overrides(cls, suffix_overrides=None):
-        """Assigns custom mappings of file suffix to style"""
-        setattr(cls, '__suffix_overrides', suffix_overrides)
-
-    @classmethod
-    def from_suffix(cls, ext):
-        """Tries to determine the style based on a file suffix"""
-        mapping = {
-            '.cc': Style.C_STYLE,
-            '.cxx': Style.C_STYLE,
-            '.cpp': Style.C_STYLE,
-            '.c': Style.C_STYLE,
-            '.hpp': Style.C_STYLE,
-            '.h': Style.C_STYLE,
-            '.hxx': Style.C_STYLE,
-            '.mm': Style.C_STYLE,
-            '.m': Style.C_STYLE,
-            '.qml': Style.C_STYLE,
-            '.java': Style.C_STYLE,
-            '.glsl': Style.C_STYLE,
-            '.frag': Style.C_STYLE,
-            '.vert': Style.C_STYLE,
-            '.rb': Style.POUND_STYLE,
-            '.py': Style.DOCSTRING_STYLE,
-            '.sh': Style.POUND_STYLE,
-            '.bash': Style.POUND_STYLE,
-            '.command': Style.POUND_STYLE,
-            '.cmake': Style.POUND_STYLE,
-            '.xml': Style.XML_STYLE,
-            '.htm': Style.XML_STYLE,
-            '.html': Style.XML_STYLE,
-            '.ui': Style.XML_STYLE,
-            '.qrc': Style.XML_STYLE,
-            '.svg': Style.XML_STYLE,
-            '.bat': Style.BATCH_STYLE,
-            '.rc': Style.SLASH_STYLE,
-            '.yml': Style.POUND_STYLE,
-            '.yaml': Style.POUND_STYLE,
-            '.lua': Style.DASH_STYLE,
-            '.rs': Style.SLASH_STYLE,
-            '.toml': Style.POUND_STYLE
-        }
-        suffix_overrides = getattr(cls, '__suffix_overrides', None)
-        if suffix_overrides and ext in suffix_overrides:  # pylint: disable=unsupported-membership-test
-            return Style[suffix_overrides[ext]]  # pylint: disable=unsubscriptable-object
-        return mapping.get(ext, None) or mapping.get(ext.lower(), Style.UNKNOWN)
-
-    @staticmethod
-    def from_name(name):
-        """Tries to determine the style based on a file name"""
-        mapping = {
-            'CMakeLists.txt': Style.POUND_STYLE,
-            'requirements.txt': Style.POUND_STYLE,
-            'Cargo.lock': Style.POUND_STYLE,
-            'Cargo.toml': Style.POUND_STYLE
-        }
-        return mapping.get(name, Style.UNKNOWN)
-
-    @staticmethod
-    def patterns(style=None):
-        """
-        Returns a list of regex and style pairs
-
-        Each regex has two match groups, one for the license and one for the remainder
-
-        :style: Use to limit the list of patterns to the given style
-        """
-        style_patterns = [
-            (Style.C_STYLE,
-                r"/\*(?P<authors>.+?)@LICENSE_HEADER_START@(?P<license>.+?)\* +@LICENSE_HEADER_END@(?:.*?)\*/(?P<body>.*)"),
-            (Style.C_STYLE,
-                r"/\*(?P<authors>.+?)@MLBA_OPEN_LICENSE_HEADER_START@(?P<license>.+?)\* +@MLBA_OPEN_LICENSE_HEADER_END@(?:.*?)\*/(?P<body>.*)"),
-            (Style.C_STYLE,
-                r"/\*(?P<authors>.+?)All rights reserved\.(?P<license>.+?)\*/(?P<body>.*)"),
-            (Style.POUND_STYLE,
-                r"#(?P<authors>.+?)@LICENSE_HEADER_START@(?P<license>.+?)# +@LICENSE_HEADER_END@(?:.*?)#\r?\n(?P<body>.*)"),
-            (Style.POUND_STYLE,
-                r"#(?P<authors>.+?)All rights reserved\.(?P<license>.+?)^(?P<body>[^#](.*)|$)"),
-            (Style.DOCSTRING_STYLE,
-                r"\"\"\"\r?\n(?P<authors>.+?)@LICENSE_HEADER_START@(?P<license>.+?)@LICENSE_HEADER_END@(?:.*?)\r?\n\"\"\"(?P<body>.*)"),
-            (Style.DOCSTRING_STYLE,
-                r"\"\"\"\r?\n(?P<authors>.+?)All rights reserved\.(?P<license>.+?)\"\"\"\r?\n(?P<body>.*)"),
-            (Style.DOCSTRING_STYLE,
-                r"#(?P<authors>.+?)@LICENSE_HEADER_START@(?P<license>.+?)# +@LICENSE_HEADER_END@(?:.*?)#\r?\n(?P<body>.*)"),
-            (Style.DOCSTRING_STYLE,
-                r"#(?P<authors>.+?)All rights reserved\.(?P<license>.+?)^(?P<body>[^#](.*)|$)"),
-            (Style.XML_STYLE,
-                r"<!--\r?\n(?P<authors>.+?)@LICENSE_HEADER_START@(?P<license>.+?)@LICENSE_HEADER_END@(?:.*?)\r?\n-->(?P<body>.*)"),
-            (Style.XML_STYLE,
-                r"<!--\r?\n(?P<authors>.+?)All rights reserved\.(?P<license>.+?)-->\r?\n(?P<body>.*)"),
-            (Style.BATCH_STYLE,
-                r"REM(?P<authors>.+?)@LICENSE_HEADER_START@(?P<license>.+?)@LICENSE_HEADER_END@(?:.*?)REM\r?\n(?!REM)(?P<body>.*)"),
-            (Style.BATCH_STYLE,
-                r"REM(?P<authors>.+?)All rights reserved\.(?P<license>.+?)^(?P<body>(?!REM)(.*)|$)"),
-            (Style.BATCH_STYLE,
-                r"::(?P<authors>.+?)@LICENSE_HEADER_START@(?P<license>.+?)@LICENSE_HEADER_END@(?:.*?)::\r?\n(?!::)(?P<body>.*)"),
-            (Style.BATCH_STYLE,
-                r"::(?P<authors>.+?)All rights reserved\.(?P<license>.+?)^(::)?\r?\n(?!::)(?P<body>.*)"),
-            (Style.SLASH_STYLE,
-                r"//(?P<authors>.+?)@LICENSE_HEADER_START@(?P<license>.+?)// +@LICENSE_HEADER_END@(?:.*?)//\r?\n(?P<body>.*)"),
-            (Style.SLASH_STYLE,
-                r"//(?P<authors>.+?)All rights reserved\.(?P<license>.+?)^(?P<body>[^/](.*)|$)"),
-            (Style.DASH_STYLE,
-                r"--(?P<authors>.+?)All rights reserved\.(?P<license>.+?)^(?P<body>[^-](.*)|$)"),
-            (Style.UNKNOWN,
-                r"(?P<authors>.+?)@LICENSE_HEADER_START@(?P<license>.+?)@LICENSE_HEADER_END@(?P<body>.*)"),
-        ]
-        if style is None:
-            style = Style.UNKNOWN
-        if style == Style.UNKNOWN:
-            # slow path, we have to try all of the patterns
-            return style_patterns
-        # fast path, move the expected style to the front
-        return sorted(style_patterns, key=lambda s: s[0] != style)
-
-    @staticmethod
-    def declarations():
-        """
-        Returns a list of document declarations retained at the first line
-
-        Each entry is a pair of matching regex and additional flags required
-        """
-        return [
-            # unicode bom marker
-            (r'\uFEFF', 0),
-            # something like '#!/usr/bin/env bash'
-            (r'^#!.+', 0),
-            # something like '# -*- coding: utf-8 -*-'
-            (r'^# -\*-.+', 0),
-            # something like '<?xml version="1.0" encoding="UTF-8" standalone="no"?>'
-            # note: use xmllint to easily validate generated output
-            (r'^<\?xml .+?\?>', 0),
-            # something like '<!DOCTYPE svg PUBLIC "-//W3C//DTD SVG 1.1//EN" "http://www.w3.org/Graphics/SVG/1.1/DTD/svg11.dtd">'
-            (r'^<!DOCTYPE .+?>', 0),
-            # beginning of batch files silencing output but only if the very first line
-            (r'^@echo off', 0)
-        ]
-
-    @staticmethod
-    def decorators(style):
-        """
-        Returns the decorator descriptions for a given style
-
-        Expect a named tuple with entries for start, prefix, end
-        as well as a pattern to strip these decorators from a line
-        """
-        Decorator = namedtuple('Decorator', 'start prefix end pattern')
-        if style == Style.C_STYLE:
-            return Decorator('/*', ' *', ' */', r' ?(?:\*) ?')
-        if style == Style.POUND_STYLE:
-            return Decorator(None, '#', None, r' ?(?:#) ?')
-        if style == Style.DOCSTRING_STYLE:
-            # the pattern will help to translate any #-style and """-style docstrings
-            return Decorator(None, '#', None, r' ?(?:#) ?')
-        if style == Style.XML_STYLE:
-            return Decorator('<!--', '', '-->', None)
-        if style == Style.BATCH_STYLE:
-            return Decorator(None, 'REM', None, r' ?(?:REM|::) ?')
-        if style == Style.SLASH_STYLE:
-            return Decorator(None, '//', None, r' ?(?://) ?')
-        if style == Style.DASH_STYLE:
-            return Decorator(None, '--', None, r' ?(?:--) ?')
-        return Decorator('', '', '', None)
-
-
-class Author:
-    """Describes an author of a file"""
-
-    def __init__(self, name: str, year_from: int = None,
-                 year_to: int = None,
-                 git_repo=None):
-        """
-        Creates a new author
-        :name: The name of the author
-        :year_from: The first year the author contributed. Will default
-                  to year_to when omitted.
-        :year_to: The last year the author contributed. Will default
-                  to DateUtils.CURRENT_YEAR when omitted.
-        """
-        self.name = name
-        self.git_repo = git_repo
-        self.name_from_git = git_repo is not None
-        if year_to is None:
-            git_author_date = os.environ.get('GIT_AUTHOR_DATE', None)
-            if self.name_from_git and git_author_date:
-                year_to = DateUtils.parse_git_date(git_author_date).year
-            else:
-                year_to = DateUtils.current_year()
-        self.year_to = year_to
-        if year_from:
-            self.year_from = year_from
-        else:
-            self.year_from = year_to
-
-    def __repr__(self) -> str:
-        return f"({self.name} {self.year_from}-{self.year_to})"
-
-    def __eq__(self, other: object) -> bool:
-        return self.name == other.name and self.year_from == other.year_from and self.year_to == other.year_to
-
-
-class GitRepo:
-    """Git repository object"""
-
-    @staticmethod
-    @functools.lru_cache(maxsize=256, typed=True)
-    def find_git_root(cwd: pathlib.Path) -> pathlib.Path:
-        """Tries to find the git root as seen from cwd"""
-        root = subprocess.check_output(
-            'git rev-parse --show-toplevel', cwd=cwd, stderr=subprocess.STDOUT, shell=True, encoding='utf-8')
-        return pathlib.Path(root.strip())
-
-    @staticmethod
-    @functools.lru_cache(maxsize=256, typed=True)
-    def author_name_from_config(cwd: pathlib.Path) -> str:
-        """Returns the author name as set via gitconfig and seen from cwd"""
-        git_author = subprocess.check_output(
-            'git config user.name', cwd=cwd, stderr=subprocess.STDOUT, shell=True, encoding='utf-8')
-        return git_author.strip()
-
-    def __init__(self, cwd=None):
-        """
-        Creates a repository object using the current working dir to determine the root
-
-        :cwd: Directory to create the repo obj from, leave None to use current wkdir
-        :throws RuntimeError: When failing to detect the root
-        """
-        try:
-            if cwd is None:
-                cwd = CW_DIR
-            self.git_root = GitRepo.find_git_root(cwd)
-        except subprocess.CalledProcessError as error:
-            raise RuntimeError(f"Not a git repo: {error.output}") from error
-
-    def author_from_config(self) -> Author:
-        """Returns the author as set via gitconfig"""
-        try:
-            return Author(name=GitRepo.author_name_from_config(self.git_root), git_repo=self)
-        except subprocess.CalledProcessError as error:
-            logging.fatal(f"Failed to fetch author using git: {error.output}")
-            return None
-
-    def author_from_history(self, filename: pathlib.Path) -> Author:
-        """Returns the author who touched the file for the last time or None"""
-        file_rel = filename.relative_to(self.git_root)
-        try:
-            author_raw = subprocess.check_output(
-                f'git log -1 --date=format:%Y --pretty=format:"%an\t%ad" -- {file_rel}',
-                cwd=self.git_root, stderr=subprocess.STDOUT, shell=True,
-                encoding='utf-8')
-            author_raw = author_raw.strip()
-            author_name, author_year = author_raw.split('\t')
-            author_year = int(author_year)
-            logging.debug(f"{file_rel} was last touched by \"{author_name}\" during {author_year}")
-            return Author(name=author_name, year_to=author_year, git_repo=self)
-        except ValueError:
-            pass
-        except subprocess.CalledProcessError:
-            pass
-        return None
-
-    def is_modified_in_tree(self, filename: pathlib.Path) -> bool:
-        """Returns true when the file has uncommited chnages in the tree"""
-        file_rel = filename.relative_to(self.git_root)
-        try:
-            diff = subprocess.check_output(
-                f'git diff --name-only HEAD -- {file_rel}',
-                cwd=self.git_root, stderr=subprocess.STDOUT, shell=True,
-                encoding='utf-8').strip()
-            if len(diff) != 0:
-                logging.debug(f"{file_rel} was just modified: {diff}")
-                return True
-        except subprocess.CalledProcessError:
-            pass
-        return False
-
-
-class License:
-    """Describes a license to be added to a header"""
-
-    def __init__(self, builtin: str = None, custom: str = None):
-        """
-        Creates a new license
-
-        Either builtin or custom need to be specified.
-
-        :builtin: One of the supported licenses, see LICENSES
-        :custom: Configures a custom license text
-        """
-        if builtin:
-            self.name = builtin
-            self.header = LICENSES.get(builtin, None)
-            if self.header is None:
-                raise TypeError(f"No such license '{builtin}'")
-            self.builtin = True
-            self.spdx = self.header in SPDX_LICENSES
-            self.header = self.header.parent.name + '/' + self.header.name
-        elif custom:
-            self.name = 'custom'
-            self.header = custom
-            self.builtin = False
-            self.spdx = False
-        else:
-            raise KeyError("Need to select a 'builtin' or provide 'custom' license")
-
-
-class Title:
-    """Describes the title to be added to a header"""
-
-    BUILTINS = ['filename']
-
-    def __init__(self, builtin: str = None, custom: str = None):
-        """
-        Creates a new title
-
-        Either builtin or custom need to be specified.
-
-        :builtin: One of the builtin titles, see BUILTINS
-        :custom: Configures a custom title text
-        """
-        self.filename = False
-        if builtin == 'filename':
-            self.filename = True
-        elif custom:
-            self.custom = custom
-        else:
-            raise KeyError("Need to select a 'builtin' or provide 'custom' title")
-
-    def get(self, file: pathlib.Path) -> str:
-        '''Determines the title for the given file'''
-        if self.filename:
-            return file.name
-        return self.custom
-
-
-class Header:
-    """Describes a header to be rendered with license and authors"""
-
-    def __init__(self, default_license, lines_after_license: int = 1):
-        """Creates a new header using given default license"""
-        loader = jinja2.FileSystemLoader(BASE_DIR)
-        self.env = jinja2.Environment(loader=loader)
-        self.template = self.env.get_template('Header.j2')
-        self.default_license = default_license
-        self.lines_after_license = lines_after_license
-
-    def render(self, title: str, authors, style: Style, company: str = None, license: str = None) -> str:
-        """
-        Renders a header to a string
-        :title: The title to put at the top of the file, usually the filename
-        :authors: A list of authors which contributed to the file
-        :style: The comment style to use when generating the header
-        :company: An optional company string to be included
-        :license: An optional license string to be used, if omitted the default_license will apply
-        """
-        if company is None:
-            company = 'the authors'
-        header = self.template.render(default_license=self.default_license,
-                                      license=license, title=title, authors=authors, company=company)
-        decorators = Style.decorators(style)
-        header = header.split('\n')
-        header = [f'{decorators.prefix} ' + h if h.strip() else decorators.prefix for h in header]
-        if decorators.start:
-            header = [decorators.start] + header
-        if decorators.end:
-            header = header + [decorators.end]
-        return '\n'.join(header) + '\n' * max(self.lines_after_license, 1)
-
-
-class ParsedHeader:
-    """A license header parsed from an existing file"""
-
-    def __init__(self, file: pathlib.PurePath = None, contents: str = None):
-        """
-        Parses a header from the given file
-
-        :file: The filepath of the header
-        :contenst: The contents of the header, if None this will be read from file
-        """
-        if contents is None:
-            with open(file, 'r', encoding='utf-8', newline='') as file_obj:
-                contents = file_obj.read()
-        if isinstance(file, str):
-            file = pathlib.Path(file)
-        # style is determined from the extension, if unknown we try a second attempt using the contents below
-        self.style = Style.from_suffix(file.suffix)
-        if self.style == Style.UNKNOWN:
-            self.style = Style.from_name(file.name)
-        # strip but remember any shebang, encoding or doctype at the beginning
-        self.decls = []
-
-        def extract_decl(pattern, contents, flags=re.MULTILINE):
-            decl = re.match(pattern, contents, flags)
-            if decl:
-                decl = decl[0]
-                self.decls.append(decl)
-                # remove the decl including the newline
-                return contents[len(decl):].lstrip()
-            return contents
-        for decl, flags in Style.declarations():
-            contents = extract_decl(decl, contents, flags)
-        # any known license is wrapped in well-known tags
-        for style, pattern in Style.patterns(self.style):
-            match = re.match(pattern, contents, re.MULTILINE | re.DOTALL)
-            if match:
-                match_style = style
-                if self.style == Style.UNKNOWN:
-                    self.style = style
-                break
-        if match:
-            # grab the matched license but remove any # or * per line prefix decorators
-            self.license = match.group('license')
-            decorators = Style.decorators(match_style)
-            if decorators.pattern:
-                self.license = re.sub(decorators.pattern, '', self.license, flags=re.MULTILINE)
-            self.license = self.license.strip('\n\r')
-            if self.license.startswith(' '):
-                # filter any leading indends
-                self.license = self.license.replace('\n ', '\n')
-            self.license = self.license.strip()
-            if self.license == "":
-                # When license is an empty string reset to None
-                self.license = None
-            self.remainder = match.group('body').strip()
-        else:
-            self.license = None
-            self.remainder = contents.strip()
-        # determine the line endings from the remainder or default to platform if none
-        if self.remainder:
-            if '\r\n' in self.remainder:
-                self.newline = '\r\n'
-            else:
-                self.newline = '\n'
-        else:
-            self.newline = os.linesep
-        # in case we have a matched header we can use its authors group to limit our search
-        # use a regex to extract existing authors, i.e. any line starting with 'Copyright'
-        if match:
-            authors_raw = match.group('authors') or contents
-        else:
-            authors_raw = contents
-        self.authors = []
-        for match in re.finditer(r" Copyright[^\d]*(?P<from>[0-9]+) *(?:- *(?P<to>[0-9]+))? *(?P<name>[^\n\r]+)", authors_raw, re.IGNORECASE):
-            args = {
-                'name': match.group('name'),
-                'year_from': int(match.group('from'))
-            }
-            try:
-                args['year_to'] = int(match.group('to'))
-            except TypeError:
-                args['year_to'] = args['year_from']
-            author = Author(**args)
-            self.authors.append(author)
-        self.authors = sorted(self.authors, key=attrgetter('year_from'))
 
 
 class Tool:
@@ -700,97 +170,6 @@ class Tool:
                 output.write(bumped)
                 return True
         return False
-
-
-class FileFilter:
-    """Helper to filter against an exclude and include list"""
-
-    @staticmethod
-    @functools.lru_cache(maxsize=256, typed=True)
-    def _glob_to_re(expr: str) -> re.Pattern:
-        """
-        Converts a globbing expression to a regex
-
-        In contrast to fnmatch it supports recursive
-        globbing as well and is cached similar to
-        fnmatch.fnmatch is optimizing repeated calls.
-        """
-        class State(enum.Enum):
-            '''Enumerates parsing state'''
-            TEXT = 0
-            SEQ = 1
-            GLOB = 2
-
-        state = State.TEXT
-        pattern = '^'
-        i, n = 0, len(expr)  # pylint: disable=invalid-name
-        while i < n:
-            # when in a sequence continue
-            # up to the terminator
-            if state == State.SEQ:
-                if expr[i] == '[':
-                    raise RuntimeError(f"Character sequences cannot be nested: {expr}")
-                if expr[i] == ']':
-                    pattern += ']'
-                    state = State.TEXT
-                else:
-                    pattern += re.escape(expr[i])
-            elif expr[i] == ']':
-                raise RuntimeError(f"Closing sequence not opened before: {expr}")
-            elif expr[i:i+2] == '[!':  # negated sequence
-                state = State.SEQ
-                pattern += '[^'
-                i += 1
-            elif expr[i] == '[':  # sequence
-                state = State.SEQ
-                pattern += '['
-            elif expr[i:i+3] == '**/':  # recursive glob
-                pattern += '(.+/)?'
-                i += 2
-            elif expr[i] == '?':  # single char
-                pattern += '[^/]'
-            elif expr[i] == '*':  # basic globbing
-                if state != State.GLOB:
-                    pattern += '[^/]*'
-                    state = State.GLOB
-                # else: Skip consecutive *
-            else:
-                pattern += re.escape(expr[i])
-                state = State.TEXT
-            i += 1
-        pattern += '$'
-        return re.compile(pattern)
-
-    @staticmethod
-    @functools.lru_cache(maxsize=256, typed=True)
-    def _to_re(pattern: str) -> re.Pattern:
-        return re.compile(pattern)
-
-    @staticmethod
-    def is_included(file_rel: str, includes, excludes) -> bool:
-        """
-        Tests if a given relative file matches includes and not excludes
-
-        :file_rel: Relative filepath to be tested
-        :includes: List of globbing expressions noting files to includes
-        :exclude: List of regular expressions noting files to exclude
-        """
-        matched = False
-        match_reason = f"Excluding '{file_rel}' - failed to match any"
-        for include in includes:
-            include = FileFilter._glob_to_re(include)
-            if re.match(include, file_rel):
-                match_reason = f"Including '{file_rel}' because of '{include.pattern}'"
-                matched = True
-                break
-        for exclude in excludes:
-            exclude = FileFilter._to_re(exclude)
-            if re.search(exclude, file_rel):
-                match_reason = f"Excluding '{file_rel}' due to '{exclude.pattern}'"
-                matched = False
-                break
-        logging.debug(match_reason)
-        return matched
 
 
 def main():
